Switch over from an XML based file format to .INI-style conf files
This change moves us to using a single file format for all configuration
throughout bluedroid. I've provided a transcoder that converts the
previous config file to the new format so settings will be preserved.

In a future revision of bluedroid, we should remove the transcoder and
drop our dependency on tinyxml2.