Fix NULL pointer dereference in bta_pan_co_tx_path.
This happens when the PAN connection terminates but the stack is
still sending data back to the system. There's a timing component
to this bug; the stack doesn't crash 100% of the time since it
depends on whether the TX path notices the PAN connection was torn
down or not before calling bta_pan_co_tx_path.