Fixed HAL layer ignore state handling
If the hci layer ever entered the ignore state,
synchronization death was immenent. This patch fixes that.
Also adds additional info print out when entering the ignored state.

Adds tests for the ignored state.

Doesn't solve the problem of why we were getting NULL back from
the buffer allocator, though. (The trigger for the ignored state.)

Bug: 18156298