Fix alarm setting for newer deadlines
If an alarm is being set and is the newest deadline
among a set of existing alarms, it would be appended
after the first existing alarm rather than prepended
in front.

This in turn meant the root alarm wasn't getting
rescheduled correctly, especially if the old
deadline was rather distant in the future.

(This was causing the bluetooth disable sequence to
fail intermittently.)