Wi-Fi HAL: Initial implementation of Ring Buffer
Ring buffer consists of discrete buffers alloced and freed as and
when required and not required. These buffers together will
operate as a ring of memory and allows write and read operations
of different sizes. Also provides facility to trigger external
callbacks when the size of ring data crosses configured threshold.

Change-Id: Ide7453dad59a1abc82381619831e268e82ef89da