wifi:Stub out support for SET_AP_WPS_P2P_IE command
The supplicant is sending SET_AP_WPS_P2P_IE driver command
which is not handled by driver. This is an API that is directly
called by Android extensions to the wpa_supplicant. As the
SET_AP_WPS_P2P_IE command is not handled, the handler function
for this command is removed in wpa_supplicant_8_lib.

Change-Id: Ia54f924b4340ccd243716b1af23d978a92823eb2