build: Add gtest as a secondary source
As a temporary solution, this CL introduces gtest as a secondary build
source. For now, gtest can be included in testonly targets as a direct
GN target dependency, by placing the gtest source under "testing/gtest"
at the root of the tree.

Moving forward, we can use solutions like gclient to pull in external
dependencies such as gtest, or we can depend on it being available on
the system.

BUG=21339022,21340407

Change-Id: I80ab3a81a53d7c29d54857a8d0cd9677fed77549