WiFi-HAL: Validate debug events with sufficient length only
Frequent updations are expected to the stats structures from
firmware. Validating the events received based on its exact size
may not be correct in this case. So, consider events as valid only
if the size of events is not less than expected.

Rx peer info event has been handled with the help flags already.
So, remove the duplicate check from parse_rx_stats()

Change-Id: I42f2a3f402684fd3de049d7d9fe011068fabb45b