Wifi-HAL: Avoid Gscan race-condition between events/commands
We are seeing a race-condition when the event handler function
is getting events from the driver and at the same time cleanup
is also happening from the command side. Due to the race-condition
we are seeing a crash. This commit addresses this by ensuring that
the cleanup of event handler happens only when wifihal cleanup is
happening and also would happen in the same context in which events
are processed.

BUG: 23277168
Change-Id: Ib0cbcbd23e2c0424d6f81597685d085d95df340d