test_vendor_lib: Clean up controller documentation and address style issues
This upload cleans up the command documentation in the controller's
header file and also fixes a few logical errors related to command
status event packets. Some style fixes are also included, such as
removing braces for single line conditionals to conform to remain
consistent with the rest of the project.

Bug: 21586676
Change-Id: Ife5e608e43c3030e76eff8ea79ce6cad0b2c10a3