HID: Remove bond for device in case Virtual unplug is ongoing
Remove bond for remote device in case virtual unplug was initiated from
local device, and ACL got disconnected while processing the VUP request.
this can happen when aaplication is sending disconnect and remove bond
in quick succession. Without this fix it was observed that device was
getting disconnected, but was not getting unpaired even when user
tried to unpair from UI.

Bug: 27852645
Change-Id: Ife1fc99ed59b13daf0bcf0dce3e0a59c7d9e87d7