QCamera2: Set sane default video sizes
Most apps do not expect resolutions greater than 1080p as the default video
size, so set the default video size for the rear camera to 1080p. For the
front-facing camera, set the default video size to 720p (as resolutions greater
than 720p on the FFC are capped to 15 FPS).

This fixes the incorrect aspect ratio when video calling in Hangouts.