Wifi-HAL: Avoid illegal memory access in wifi_set_packet_filter fn.
In requestResponse(), nl message is getting freed after receiving
a response for the nl message, and wifi_hal is trying to fill the same
message again in wifi_set_packet_filter() for next fragment of
data. This is causing an invalid memory access. To avoid this,
destroy the object after sending each fragment to driver and
allocate a fresh object which in turn creates nl message and
initializes it.

BUG: 27595799
Change-Id: I654203a55206a5efe820146bde43c20b76b55880