Wifi-HAL: Extract wmsg length from nlmsg_hdr with ntohs()
Driver swaps the wmsg length based on endianness for
ANI_NL_MSG_PUMAC messages. So, use ntohs() to get correct length
of wmsg. The reason this swap is done in Wi-Fi HAL is to co work
with the already existing tools that already does the same while
operating with the host driver.