WiFi-HAL: Support timestamp sync in firmware diag message
As part of timestamp synchronization firmware events, logs and F3
message formats have changed which would result in change in
processing of the same. This commit will address the change in
Wifi-Hal catering to the new format, at the same time maintaining
backward compatibility.
At least FW 4.3-00089 is required for working.

BUG: 29762703