WiFi-HAL: Avoid race condition in RSSI monitoring commands/events
There is a possiblility while handling the wlan driver event
in the handleEvent of the RssiMonitor, the RssiMonitor stop
is also triggered. The stop monitor cleans RssiMonitor command
instance. In this case the handleEvent will crash as the object
is already deleted. To avoid this scenario moving the handler
creation under hal_info and adding checks to ensure that handler
is enabled/disabled prior to starting and stopping the monitor
command.

BUG: 29936203
Change-Id: Id0ab73313a03510f0b15b101e381112c4ace4aef