Add security record for devices found through LE batch scan
Currently, starting an LE batch scan are not added to the stack internal
device database, which means that the address type and device type
needed for a connection are not available.

This patch ensures that devices found during an LE scan are added to the
device database to ensure connection attempts will succeed.

Bug: 31309376
Change-Id: Ic52d7eaa03d17dcbbd848af9140038bc84ea3bd0