Make bta_gattc_listen only start listening
The bta_gattc_listen method is used only to start advertising on devices
not using Vendor-Specific Callbacks. It should not notify the client
calling it of all connected devices.

It is also unnecessary to add the gatt client as listening for background
connections. No such thing is done for multi advertising.

Bug: 30622771
Bug: 24099160
Test: code is still compiling
Change-Id: Ia11a247b8416f77543087cd3f18331fcfe7bd207