[AWARE] Fix the Nan register handler command
Nan SetCallBackHandler is currently registering NAN CB and NDP
CB. During Wifi/NAN ON/OFF test sequence the mSubcmd being a
member variable is getting over written when an event is received
from the driver. The NDP CB is getting registered twice in this
case and the NAN subCmd handler is never registered.

BUG: 29105100
Change-Id: I7b04b5460c3ec24e32462f27f4437c7a0c4eb948