WiFi-Hal: Support new header of size 16-bytes for fate stats
Host driver sends the stats with 15th-bit set in flags field of
wh_pktlog_hdr_t header if the structure size is 16-bytes.
Offset for the fates stats from header by size of new
structure (16-bytes) if the flag is set. Otherwise offset by the
size of existing structure (12 bytes).

Change-Id: I5941c3ac5f525f36291ebee8eff01309dee11cc8