Snap: Never reuse an old FocusOverlayManager instance
Reusing a previous FocusOverlayManager instance can cause it to be
populated with stale data after switching between the front and rear
cameras in certain patterns. When mFocusManager uses stale data, incorrect
focus modes can be used and create problems. One such issue is that the
rear camera can get stuck using only fixed focus, as that's the only focus
mode that the front camera supports.

In order to make sure that all instances of FocusOverlayManager actually
correspond to the camera in use, create a fresh instance every time a
camera is opened.