Lower the BLE connection supervision timeout to 5s.
The current default of 20 seconds is too high especially considering
that BLE tries to optimize (re)connection times. A connection
supervision timeout of 20 seconds essentially means that an initiator
of the connection could be in the connecting state for 20 seconds
before giving up on that connection attempt.

Checking other platforms such as iOS for default connection supervision
timeout reveals that they use sub-second (750ms) values.

Change-Id: I6ff7a20ceab0df8986191dd73676cadbe23a175e