Fix SIGBUS error in BT stack when pairing with ARM boards
Bluetooth stack crashes with a SIGBUS memory unaligned error since the
toolchain optimizes memcpy to generate register instructions which are
unaligned in memory. Similar issue was seen earlier in
https://android-review.googlesource.com/#/c/181713/
which was solved by casting pointers to (void *). That fix misses edge
cases where the destination pointer needs to be casted as well.

Change-Id: I23cc16142deaf3dc1fff91588d736abdfcb3bfde
Fixes: 35154933
Test: Verified by building and pairing with RPI3/NXP Arm boards.