Close leftover Audio-to-Bluetooth HAL connections before reopening
If the Audio-to-Bluetooth HAL reopens a connection on the Data or
Control channel without closing first the old one, the UIPC select(2)
loop gets confused and could spin with 100% CPU usage.
Now we check for channels that are reopened, and explicitly close
them before reopening.

Also, fix a check whether a file descriptor is valid: 0 is a valid
file descriptor.

Test: A2DP testing with a headset: connect/play/stop
Bug: 35394629
Change-Id: I7ae988ac77f25da56e59e347363e7c473e7f69a9